moved item data to api folder
