--- components/data/itemData.js
@@ -1,79 +1,0 @@
-import client from '../../utils/client';
-
-const endpoint = client.databaseURL;
-
-const getItems = () => new Promise((resolve, reject) => {
-  fetch(`${endpoint}/item.json`, {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json'
-    }
-  })
-    .then((response) => response.json())
-    .then((data) => {
-      if (data) {
-        resolve(Object.values(data));
-      } else {
-        resolve([]);
-      }
-    })
-    .catch(reject);
-});
-
-const createItem = (payload) => new Promise((resolve, reject) => {
-  fetch(`${endpoint}/item.json`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(payload),
-  })
-    .then((response) => response.json())
-    .then((data) => resolve(data))
-    .catch(reject);
-});
-
-const updateItem = (payload) => new Promise((resolve, reject) => {
-  fetch(`${endpoint}/item/${payload.firebaseKey}.json`, {
-    method: 'PATCH',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(payload),
-  })
-    .then((response) => response.json())
-    .then(resolve)
-    .catch(reject);
-});
-
-const deleteItem = (firebaseKey) => new Promise((resolve, reject) => {
-  fetch(`${endpoint}/item/${firebaseKey}.json`, {
-    method: 'DELETE',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  })
-    .then((response) => response.json())
-    .then((data) => resolve(data))
-    .catch(reject);
-});
-
-const getSingleItem = (firebaseKey) => new Promise((resolve, reject) => {
-  fetch(`${endpoint}/item/${firebaseKey}.json`, {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  })
-    .then((response) => response.json())
-    .then((data) => resolve(data))
-    .catch(reject);
-});
-
-export {
-  getItems,
-  deleteItem,
-  createItem,
-  updateItem,
-  getSingleItem
-};
